+++ ts-rs/src/lib.rs
@@ -44,9 +44,7 @@
 //! ```
 //!
 //! ```rust
-//! use ts_rs::TS;
-//!
-//! #[derive(TS)]
+//! #[derive(ts_rs::TS)]
 //! #[ts(export)]
 //! struct User {
 //!     user_id: i32,
@@ -55,8 +53,7 @@
 //! }
 //! ```
 //!
-//! When running `cargo test` or `cargo test export_bindings`, the TypeScript bindings will be exported to the file `bindings/User.ts`
-//! and will contain the following code:
+//! When running `cargo test` or `cargo test export_bindings`, the following TypeScript type will be exported to `bindings/User.ts`:
 //!
 //! ```ts
 //! export type User = { user_id: number, first_name: string, last_name: string, };
@@ -65,70 +62,67 @@
 //! ## Features
 //! - generate type declarations from rust structs
 //! - generate union declarations from rust enums
-//! - inline types
-//! - flatten structs/types
+//! - works with generic types
+//! - compatible with serde
 //! - generate necessary imports when exporting to multiple files
-//! - serde compatibility
-//! - generic types
-//! - support for ESM imports
+//! - precise control over generated types
 //!
-//! ## cargo features
-//! | **Feature**        | **Description**                                                                                                                                                                                           |
-//! |:-------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-//! | serde-compat       | **Enabled by default** <br/>See the *"serde compatibility"* section below for more information.                                                                                                           |
-//! | format             | Enables formatting of the generated TypeScript bindings. <br/>Currently, this unfortunately adds quite a few dependencies.                                                                                |
-//! | no-serde-warnings  | By default, warnings are printed during build if unsupported serde attributes are encountered. <br/>Enabling this feature silences these warnings.                                                        |
-//! | serde-json-impl    | Implement `TS` for types from *serde_json*                                                                                                                                                                |
-//! | chrono-impl        | Implement `TS` for types from *chrono*                                                                                                                                                                    |
-//! | bigdecimal-impl    | Implement `TS` for types from *bigdecimal*                                                                                                                                                                |
-//! | url-impl           | Implement `TS` for types from *url*                                                                                                                                                                       |
-//! | uuid-impl          | Implement `TS` for types from *uuid*                                                                                                                                                                      |
-//! | bson-uuid-impl     | Implement `TS` for *bson::oid::ObjectId* and *bson::uuid*                                                                                                                                                 |
-//! | bytes-impl         | Implement `TS` for types from *bytes*                                                                                                                                                                     |
-//! | indexmap-impl      | Implement `TS` for types from *indexmap*                                                                                                                                                                  |
-//! | ordered-float-impl | Implement `TS` for types from *ordered_float*                                                                                                                                                             |
-//! | heapless-impl      | Implement `TS` for types from *heapless*                                                                                                                                                                  |
-//! | semver-impl        | Implement `TS` for types from *semver*                                                                                                                                                                    |
-//! | smol_str-impl      | Implement `TS` for types from *smol_str*                                                                                                                                                                  |
-//! | tokio-impl         | Implement `TS` for types from *tokio*                                                                                                                                                                     |
-//! | jiff-impl          | Implement `TS` for types from *jiff*                                                                                                                                                                      |
-//! | arrayvec-impl      | Implement `TS` for types from *arrayvec*                                                                                                                                                                  |
+//! If there's a type you're dealing with which doesn't implement `TS`, you can use either
+//! `#[ts(as = "..")]` or `#[ts(type = "..")]`, enable the appropriate cargo feature, or open a PR.
 //!
-//! <br/>
-//!
-//! If there's a type you're dealing with which doesn't implement `TS`, use either
-//! `#[ts(as = "..")]` or `#[ts(type = "..")]`, or open a PR.
-//!
-//! ## `serde` compatability
-//! With the `serde-compat` feature (enabled by default), serde attributes can be parsed for enums and structs.
-//! Supported serde attributes:
-//! - `rename`
-//! - `rename-all`
-//! - `rename-all-fields`
-//! - `tag`
-//! - `content`
-//! - `untagged`
-//! - `skip`
-//! - `skip_serializing`
-//! - `skip_serializing_if`
-//! - `flatten`
-//! - `default`
-//!
-//! Note: `skip_serializing` and `skip_serializing_if` only have an effect when used together with
-//! `#[serde(default)]`.
-//!
-//! Note: `skip_deserializing` is ignored. If you wish to exclude a field
-//! from the generated type, but cannot use `#[serde(skip)]`, use `#[ts(skip)]` instead.
-//!
-//! When ts-rs encounters an unsupported serde attribute, a warning is emitted, unless the feature `no-serde-warnings` is enabled.
-//!
-//! ## Environment variables
+//! ## Configuration
+//! When using `#[ts(export)]` on a type, `ts-rs` generates a test which writes the bindings for it to disk.\
+//! The following environment variables may be set to configure *how* and *where*:   
 //! | Variable                 | Description                                                         | Default      |
 //! |--------------------------|---------------------------------------------------------------------|--------------|
 //! | `TS_RS_EXPORT_DIR`       | Base directory into which bindings will be exported                 | `./bindings` |
 //! | `TS_RS_IMPORT_EXTENSION` | File extension used in `import` statements                          | *none*       |
 //! | `TS_RS_LARGE_INT`        | Binding used for large integer types (`i64`, `u64`, `i128`, `u128`) | `bigint`     |
 //!
+//! We recommend putting this configuration in the project's [config.toml](https://doc.rust-lang.org/cargo/reference/config.html#env) to make it persistent:
+//! ```toml
+//! # <project-root>/.cargo/config.toml
+//! [env]
+//! TS_RS_EXPORT_DIR = { value = "bindings", relative = true }
+//! TS_RS_LARGE_INT = "number"
+//! ```
+//!
+//! To export bindings programmatically without the use of tests, `TS::export_all`, `TS::export`, and `TS::export_to_string` can be used instead.
+//!
+//! ## Serde Compatibility
+//! With the `serde-compat` feature (enabled by default), serde attributes are parsed for enums and structs.\
+//! Supported serde attributes: `rename`, `rename-all`, `rename-all-fields`, `tag`, `content`, `untagged`, `skip`, `skip_serializing`, `skip_serializing_if`, `flatten`, `default`
+//!
+//! **Note**: `skip_serializing` and `skip_serializing_if` only have an effect when used together with
+//! `#[serde(default)]`. This ensures that the generated type is correct for both serialization and deserialization.
+//!
+//! **Note**: `skip_deserializing` is ignored. If you wish to exclude a field
+//! from the generated type, but cannot use `#[serde(skip)]`, use `#[ts(skip)]` instead.
+//!
+//! When ts-rs encounters an unsupported serde attribute, a warning is emitted, unless the feature `no-serde-warnings` is enabled.\
+//! We are currently waiting for [#54140](https://github.com/rust-lang/rust/issues/54140), which will improve the ergonomics arund these diagnostics.
+//!
+//! ## Cargo Features
+//! | **Feature**        | **Description**                                                                                                                                     |
+//! |:-------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------|
+//! | serde-compat       | **Enabled by default** <br/>See the *"serde compatibility"* section below for more information.                                                     |
+//! | format             | Enables formatting of the generated TypeScript bindings. <br/>Currently, this unfortunately adds quite a few dependencies.                          |
+//! | no-serde-warnings  | By default, warnings are printed during build if unsupported serde attributes are encountered. <br/>Enabling this feature silences these warnings.  |
+//! | serde-json-impl    | Implement `TS` for types from *serde_json*                                                                                                          |
+//! | chrono-impl        | Implement `TS` for types from *chrono*                                                                                                              |
+//! | bigdecimal-impl    | Implement `TS` for types from *bigdecimal*                                                                                                          |
+//! | url-impl           | Implement `TS` for types from *url*                                                                                                                 |
+//! | uuid-impl          | Implement `TS` for types from *uuid*                                                                                                                |
+//! | bson-uuid-impl     | Implement `TS` for *bson::oid::ObjectId* and *bson::uuid*                                                                                           |
+//! | bytes-impl         | Implement `TS` for types from *bytes*                                                                                                               |
+//! | indexmap-impl      | Implement `TS` for types from *indexmap*                                                                                                            |
+//! | ordered-float-impl | Implement `TS` for types from *ordered_float*                                                                                                       |
+//! | heapless-impl      | Implement `TS` for types from *heapless*                                                                                                            |
+//! | semver-impl        | Implement `TS` for types from *semver*                                                                                                              |
+//! | smol_str-impl      | Implement `TS` for types from *smol_str*                                                                                                            |
+//! | tokio-impl         | Implement `TS` for types from *tokio*                                                                                                               |
+//! | jiff-impl          | Implement `TS` for types from *jiff*                                                                                                                |
+//! | arrayvec-impl      | Implement `TS` for types from *arrayvec*                                                                                                            |
 //!
 //! ## Contributing
 //! Contributions are always welcome!
@@ -148,7 +142,6 @@ use std::{
     },
     ops::{Range, RangeInclusive},
     path::{Path, PathBuf},
-    sync::OnceLock,
 };
 
 pub use ts_rs_macros::TS;
@@ -183,19 +176,12 @@ mod tokio;
 /// `#[ts(export_to = "...")]`. By default, the filename will be derived from the name of the type.
 ///
 /// If, for some reason, you need to do this during runtime or cannot use `#[ts(export)]`, bindings
-/// can be exported manually:
-///
-/// | Function              | Includes Dependencies | To                 |
-/// |-----------------------|-----------------------|--------------------|
-/// | [`TS::export`]        | ❌                    | `TS_RS_EXPORT_DIR` |
-/// | [`TS::export_all`]    | ✔️                    | `TS_RS_EXPORT_DIR` |
-/// | [`TS::export_all_to`] | ✔️                    | _custom_           |
+/// can be exported manually using [`TS::export_all`], [`TS::export`], or [`TS::export_to_string`].
 ///
 /// ### serde compatibility
-/// By default, the feature `serde-compat` is enabled.
-/// ts-rs then parses serde attributes and adjusts the generated typescript bindings accordingly.
+/// With the `serde-compat` feature enabled (default), ts-rs parses serde attributes and adjusts the generated typescript bindings accordingly.  
 /// Not all serde attributes are supported yet - if you use an unsupported attribute, you'll see a
-/// warning.
+/// warning. These warnings can be disabled by enabling the `no-serde-warnings` cargo feature.
 ///
 /// ### container attributes
 /// attributes applicable for both structs and enums
@@ -417,11 +403,8 @@ pub trait TS {
     ///     type WithoutGenerics = GenericType<ts_rs::Dummy, ts_rs::Dummy>;
     ///     type OptionInnerType = Self;
     ///     // ...
-    ///     # fn decl() -> String { todo!() }
-    ///     # fn decl_concrete() -> String { todo!() }
-    ///     # fn name() -> String { todo!() }
-    ///     # fn inline() -> String { todo!() }
-    ///     # fn inline_flattened() -> String { todo!() }
+    ///     # fn name(_: &ts_rs::Config) -> String { todo!() }
+    ///     # fn inline(_: &ts_rs::Config) -> String { todo!() }
     /// }
     /// ```
     type WithoutGenerics: TS + ?Sized;
@@ -433,6 +416,9 @@ pub trait TS {
     #[doc(hidden)]
     const IS_OPTION: bool = false;
 
+    #[doc(hidden)]
+    const IS_ENUM: bool = false;
+
     /// JSDoc comment to describe this type in TypeScript - when `TS` is derived, docs are
     /// automatically read from your doc comments or `#[doc = ".."]` attributes
     fn docs() -> Option<String> {
@@ -440,9 +426,9 @@ pub trait TS {
     }
 
     /// Identifier of this type, excluding generic parameters.
-    fn ident() -> String {
+    fn ident(cfg: &Config) -> String {
         // by default, fall back to `TS::name()`.
-        let name = <Self as crate::TS>::name();
+        let name = <Self as crate::TS>::name(cfg);
 
         match name.find('<') {
             Some(i) => name[..i].to_owned(),
@@ -457,23 +443,29 @@ pub trait TS {
     /// placeholders, resulting in a generic typescript definition.
     /// Both `SomeType::<i32>::decl()` and `SomeType::<String>::decl()` will therefore result in
     /// the same TypeScript declaration `type SomeType<A> = ...`.
-    fn decl() -> String;
+    fn decl(cfg: &Config) -> String {
+        panic!("{} cannot be declared", Self::name(cfg))
+    }
 
     /// Declaration of this type using the supplied generic arguments.
     /// The resulting TypeScript definition will not be generic. For that, see `TS::decl()`.
     /// If this type is not generic, then this function is equivalent to `TS::decl()`.
-    fn decl_concrete() -> String;
+    fn decl_concrete(cfg: &Config) -> String {
+        panic!("{} cannot be declared", Self::name(cfg))
+    }
 
     /// Name of this type in TypeScript, including generic parameters
-    fn name() -> String;
+    fn name(cfg: &Config) -> String;
 
     /// Formats this types definition in TypeScript, e.g `{ user_id: number }`.
     /// This function will panic if the type cannot be inlined.
-    fn inline() -> String;
+    fn inline(cfg: &Config) -> String;
 
     /// Flatten a type declaration.
     /// This function will panic if the type cannot be flattened.
-    fn inline_flattened() -> String;
+    fn inline_flattened(cfg: &Config) -> String {
+        panic!("{} cannot be flattened", Self::name(cfg))
+    }
 
     /// Iterates over all dependency of this type.
     fn visit_dependencies(_: &mut impl TypeVisitor)
@@ -490,21 +482,22 @@ pub trait TS {
     }
 
     /// Resolves all dependencies of this type recursively.
-    fn dependencies() -> Vec<Dependency>
+    fn dependencies(cfg: &Config) -> Vec<Dependency>
     where
         Self: 'static,
     {
-        let mut deps: Vec<Dependency> = vec![];
-        struct Visit<'a>(&'a mut Vec<Dependency>);
+        struct Visit<'a>(&'a Config, &'a mut Vec<Dependency>);
         impl TypeVisitor for Visit<'_> {
             fn visit<T: TS + 'static + ?Sized>(&mut self) {
-                if let Some(dep) = Dependency::from_ty::<T>() {
-                    self.0.push(dep);
+                let Visit(cfg, deps) = self;
+                if let Some(dep) = Dependency::from_ty::<T>(cfg) {
+                    deps.push(dep);
                 }
             }
         }
-        <Self as crate::TS>::visit_dependencies(&mut Visit(&mut deps));
 
+        let mut deps: Vec<Dependency> = vec![];
+        Self::visit_dependencies(&mut Visit(cfg, &mut deps));
         deps
     }
 
@@ -516,24 +509,18 @@ pub trait TS {
     /// exported automatically whenever `cargo test` is run.
     /// In that case, there is no need to manually call this function.
     ///
-    /// # Target Directory
-    /// The target directory to which the type will be exported may be changed by setting the
-    /// `TS_RS_EXPORT_DIR` environment variable. By default, `./bindings` will be used.
-    ///
-    /// To specify a target directory manually, use [`TS::export_all_to`], which also exports all
-    /// dependencies.
-    ///
     /// To alter the filename or path of the type within the target directory,
     /// use `#[ts(export_to = "...")]`.
-    fn export() -> Result<(), ExportError>
+    fn export(cfg: &Config) -> Result<(), ExportError>
     where
         Self: 'static,
     {
-        let path = <Self as crate::TS>::default_output_path()
+        let relative_path = Self::output_path()
             .ok_or_else(std::any::type_name::<Self>)
             .map_err(ExportError::CannotBeExported)?;
+        let path = cfg.export_dir.join(relative_path);
 
-        export::export_to::<Self, _>(path)
+        export::export_to::<Self, _>(cfg, path)
     }
 
     /// Manually export this type to the filesystem, together with all of its dependencies.
@@ -544,39 +531,13 @@ pub trait TS {
     /// exported automatically whenever `cargo test` is run.
     /// In that case, there is no need to manually call this function.
     ///
-    /// # Target Directory
-    /// The target directory to which the types will be exported may be changed by setting the
-    /// `TS_RS_EXPORT_DIR` environment variable. By default, `./bindings` will be used.
-    ///
-    /// To specify a target directory manually, use [`TS::export_all_to`].
-    ///
     /// To alter the filenames or paths of the types within the target directory,
     /// use `#[ts(export_to = "...")]`.
-    fn export_all() -> Result<(), ExportError>
+    fn export_all(cfg: &Config) -> Result<(), ExportError>
     where
         Self: 'static,
     {
-        export::export_all_into::<Self>(&*export::default_out_dir())
-    }
-
-    /// Manually export this type into the given directory, together with all of its dependencies.
-    /// To export only this type, without its dependencies, use [`TS::export`].
-    ///
-    /// Unlike [`TS::export_all`], this function disregards `TS_RS_EXPORT_DIR`, using the provided
-    /// directory instead.
-    ///
-    /// To alter the filenames or paths of the types within the target directory,
-    /// use `#[ts(export_to = "...")]`.
-    ///
-    /// # Automatic Exporting
-    /// Types annotated with `#[ts(export)]`, together with all of their dependencies, will be
-    /// exported automatically whenever `cargo test` is run.
-    /// In that case, there is no need to manually call this function.
-    fn export_all_to(out_dir: impl AsRef<Path>) -> Result<(), ExportError>
-    where
-        Self: 'static,
-    {
-        export::export_all_into::<Self>(out_dir)
+        export::export_all_into::<Self>(cfg)
     }
 
     /// Manually generate bindings for this type, returning a [`String`].
@@ -586,45 +547,23 @@ pub trait TS {
     /// Types annotated with `#[ts(export)]`, together with all of their dependencies, will be
     /// exported automatically whenever `cargo test` is run.
     /// In that case, there is no need to manually call this function.
-    fn export_to_string() -> Result<String, ExportError>
+    fn export_to_string(cfg: &Config) -> Result<String, ExportError>
     where
         Self: 'static,
     {
-        export::export_to_string::<Self>()
+        export::export_to_string::<Self>(cfg)
     }
 
-    /// Returns the output path to where `T` should be exported.
-    /// The returned path does _not_ include the base directory from `TS_RS_EXPORT_DIR`.
-    ///
-    /// To get the output path containing `TS_RS_EXPORT_DIR`, use [`TS::default_output_path`].
+    /// Returns the output path to where `T` should be exported, relative to the output directory.
+    /// The returned path does _not_ include any base directory.
     ///
     /// When deriving `TS`, the output path can be altered using `#[ts(export_to = "...")]`.
     /// See the documentation of [`TS`] for more details.
-    ///
-    /// The output of this function depends on the environment variable `TS_RS_EXPORT_DIR`, which is
-    /// used as base directory. If it is not set, `./bindings` is used as default directory.
     ///
     /// If `T` cannot be exported (e.g because it's a primitive type), this function will return
     /// `None`.
     fn output_path() -> Option<PathBuf> {
         None
-    }
-
-    /// Returns the output path to where `T` should be exported.
-    ///
-    /// The output of this function depends on the environment variable `TS_RS_EXPORT_DIR`, which is
-    /// used as base directory. If it is not set, `./bindings` is used as default directory.
-    ///
-    /// To get the output path relative to `TS_RS_EXPORT_DIR` and without reading the environment
-    /// variable, use [`TS::output_path`].
-    ///
-    /// When deriving `TS`, the output path can be altered using `#[ts(export_to = "...")]`.
-    /// See the documentation of [`TS`] for more details.
-    ///
-    /// If `T` cannot be exported (e.g because it's a primitive type), this function will return
-    /// `None`.
-    fn default_output_path() -> Option<PathBuf> {
-        Some(export::default_out_dir().join(<Self as crate::TS>::output_path()?))
     }
 }
 
@@ -654,13 +593,144 @@ impl Dependency {
     /// Constructs a [`Dependency`] from the given type `T`.
     /// If `T` is not exportable (meaning `T::EXPORT_TO` is `None`), this function will return
     /// `None`
-    pub fn from_ty<T: TS + 'static + ?Sized>() -> Option<Self> {
+    pub fn from_ty<T: TS + 'static + ?Sized>(cfg: &Config) -> Option<Self> {
         let output_path = <T as crate::TS>::output_path()?;
         Some(Dependency {
             type_id: TypeId::of::<T>(),
-            ts_name: <T as crate::TS>::ident(),
+            ts_name: <T as crate::TS>::ident(cfg),
             output_path,
         })
+    }
+}
+
+/// Configuration that affects the generation of TypeScript bindings and how they are exported.  
+pub struct Config {
+    // TS_RS_LARGE_INT
+    large_int_type: String,
+    // TS_RS_USE_V11_HASHMAP
+    use_v11_hashmap: bool,
+    // TS_RS_EXPORT_DIR
+    export_dir: PathBuf,
+    // TS_RS_IMPORT_EXTENSION
+    import_extension: Option<String>,
+    array_tuple_limit: usize,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            large_int_type: "bigint".to_owned(),
+            use_v11_hashmap: false,
+            export_dir: "./bindings".into(),
+            import_extension: None,
+            array_tuple_limit: 64,
+        }
+    }
+}
+
+impl Config {
+    /// Creates a new `Config` with default values.  
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Creates a new `Config` with values read from environment variables.
+    ///
+    /// | Variable                 | Description                                                         | Default      |
+    /// |--------------------------|---------------------------------------------------------------------|--------------|
+    /// | `TS_RS_EXPORT_DIR`       | Base directory into which bindings will be exported                 | `./bindings` |
+    /// | `TS_RS_IMPORT_EXTENSION` | File extension used in `import` statements                          | *none*       |
+    /// | `TS_RS_LARGE_INT`        | Binding used for large integer types (`i64`, `u64`, `i128`, `u128`) | `bigint`     |
+    pub fn from_env() -> Self {
+        let mut cfg = Self::default();
+
+        if let Ok(ty) = std::env::var("TS_RS_LARGE_INT") {
+            cfg = cfg.with_large_int(ty);
+        }
+
+        if let Ok(dir) = std::env::var("TS_RS_EXPORT_DIR") {
+            cfg = cfg.with_out_dir(dir);
+        }
+
+        if let Ok(ext) = std::env::var("TS_RS_IMPORT_EXTENSION") {
+            if !ext.trim().is_empty() {
+                cfg = cfg.with_import_extension(Some(ext));
+            }
+        }
+
+        #[allow(deprecated)]
+        if let Ok("1" | "true" | "on" | "yes") = std::env::var("TS_RS_USE_V11_HASHMAP").as_deref() {
+            cfg = cfg.with_v11_hashmap();
+        }
+
+        cfg
+    }
+
+    /// Sets the TypeScript type used to represent large integers.
+    /// Here, "large" refers to integers that can not be losslessly stored using the 64-bit "binary64" IEEE 754 float format used by JavaScript.  
+    /// Those include `u64`, `i64, `u128`, and `i128`.
+    ///
+    /// Default: `"bigint"`
+    pub fn with_large_int(mut self, ty: impl Into<String>) -> Self {
+        self.large_int_type = ty.into();
+        self
+    }
+
+    /// Returns the TypeScript type used to represent large integers.
+    pub fn large_int(&self) -> &str {
+        &self.large_int_type
+    }
+
+    /// When enabled, `HashMap<K, V>` and similar types will always be translated to `{ [key in K]?: V }`.  
+    /// Normally, with this option disabled, `{ [key in K]: V }` is generated instead, unless the key `K` is an enum.  
+    /// This option is only intended to aid migration and will be removed in a future release.
+    ///
+    /// Default: disabled
+    #[deprecated = "this option is merely meant to aid migration to v12 and will be removed in a future release"]
+    pub fn with_v11_hashmap(mut self) -> Self {
+        self.use_v11_hashmap = true;
+        self
+    }
+
+    /// Sets the output directory into which bindings will be exported.  
+    /// This affects `TS::export`, `TS::export_all`, and the automatic export of types annotated with `#[ts(export)]` when `cargo test` is run.
+    ///
+    /// Default: `./bindings`
+    pub fn with_out_dir(mut self, dir: impl Into<PathBuf>) -> Self {
+        self.export_dir = dir.into();
+        self
+    }
+
+    /// Returns the output directory into which bindings will be exported.
+    pub fn out_dir(&self) -> &Path {
+        &self.export_dir
+    }
+
+    /// Sets the file extension used for `import` statements in generated TypeScript files.  
+    ///
+    /// Default: `None`
+    pub fn with_import_extension(mut self, ext: Option<impl Into<String>>) -> Self {
+        self.import_extension = ext.map(Into::into);
+        self
+    }
+
+    /// Returns the file extension used for `import` statements in generated TypeScript files.  
+    pub fn import_extension(&self) -> Option<&str> {
+        self.import_extension.as_deref()
+    }
+
+    /// Sets the maximum size of arrays (`[T; N]`) up to which they are treated as TypeScript tuples (`[T, T, ...]`).  
+    /// Arrays beyond this size will instead result in a TypeScript array (`Array<T>`).
+    ///
+    /// Default: `64`
+    pub fn with_array_tuple_limit(mut self, limit: usize) -> Self {
+        self.array_tuple_limit = limit;
+        self
+    }
+
+    /// Returns the maximum size of arrays (`[T; N]`) up to which they are treated as TypeScript tuples (`[T, T, ...]`).  
+    pub fn array_tuple_limit(&self) -> usize {
+        self.array_tuple_limit
     }
 }
 
@@ -684,24 +754,34 @@ macro_rules! impl_primitives {
         impl TS for $ty {
             type WithoutGenerics = Self;
             type OptionInnerType = Self;
-            fn name() -> String { String::from($l) }
-            fn inline() -> String { <Self as $crate::TS>::name() }
-            fn inline_flattened() -> String { panic!("{} cannot be flattened", <Self as $crate::TS>::name()) }
-            fn decl() -> String { panic!("{} cannot be declared", <Self as $crate::TS>::name()) }
-            fn decl_concrete() -> String { panic!("{} cannot be declared", <Self as $crate::TS>::name()) }
+            fn name(_: &$crate::Config) -> String { String::from($l) }
+            fn inline(cfg: &$crate::Config) -> String { <Self as $crate::TS>::name(cfg) }
         }
     )*)* };
 }
+
+// generate impls for big integers
+macro_rules! impl_large_integers {
+    ($($ty:ty),*) => { $(
+        impl TS for $ty {
+            type WithoutGenerics = Self;
+            type OptionInnerType = Self;
+            fn name(cfg: &$crate::Config) -> String { cfg.large_int_type.clone() }
+            fn inline(cfg: &$crate::Config) -> String { <Self as $crate::TS>::name(cfg) }
+        }
+    )* };
+}
+
 // generate impls for tuples
 macro_rules! impl_tuples {
     ( impl $($i:ident),* ) => {
         impl<$($i: TS),*> TS for ($($i,)*) {
             type WithoutGenerics = (Dummy, );
             type OptionInnerType = Self;
-            fn name() -> String {
-                format!("[{}]", [$(<$i as $crate::TS>::name()),*].join(", "))
+            fn name(cfg: &$crate::Config) -> String {
+                format!("[{}]", [$(<$i as $crate::TS>::name(cfg)),*].join(", "))
             }
-            fn inline() -> String {
+            fn inline(_: &$crate::Config) -> String {
                 panic!("tuple cannot be inlined!");
             }
             fn visit_generics(v: &mut impl TypeVisitor)
@@ -713,9 +793,9 @@ macro_rules! impl_tuples {
                     <$i as $crate::TS>::visit_generics(v);
                 )*
             }
-            fn inline_flattened() -> String { panic!("tuple cannot be flattened") }
-            fn decl() -> String { panic!("tuple cannot be declared") }
-            fn decl_concrete() -> String { panic!("tuple cannot be declared") }
+            fn inline_flattened(_: &$crate::Config) -> String { panic!("tuple cannot be flattened") }
+            fn decl(_: &$crate::Config) -> String { panic!("tuple cannot be declared") }
+            fn decl_concrete(_: &$crate::Config) -> String { panic!("tuple cannot be declared") }
         }
     };
     ( $i2:ident $(, $i:ident)* ) => {
@@ -731,9 +811,9 @@ macro_rules! impl_wrapper {
         $($t)* {
             type WithoutGenerics = Self;
             type OptionInnerType = Self;
-            fn name() -> String { <T as $crate::TS>::name() }
-            fn inline() -> String { <T as $crate::TS>::inline() }
-            fn inline_flattened() -> String { <T as $crate::TS>::inline_flattened() }
+            fn name(cfg: &$crate::Config) -> String { <T as $crate::TS>::name(cfg) }
+            fn inline(cfg: &$crate::Config) -> String { <T as $crate::TS>::inline(cfg) }
+            fn inline_flattened(cfg: &$crate::Config) -> String { <T as $crate::TS>::inline_flattened(cfg) }
             fn visit_dependencies(v: &mut impl TypeVisitor)
             where
                 Self: 'static,
@@ -748,8 +828,8 @@ macro_rules! impl_wrapper {
                 <T as $crate::TS>::visit_generics(v);
                 v.visit::<T>();
             }
-            fn decl() -> String { panic!("wrapper type cannot be declared") }
-            fn decl_concrete() -> String { panic!("wrapper type cannot be declared") }
+            fn decl(_: &$crate::Config) -> String { panic!("wrapper type cannot be declared") }
+            fn decl_concrete(_: &$crate::Config) -> String { panic!("wrapper type cannot be declared") }
         }
     };
 }
@@ -760,10 +840,10 @@ macro_rules! impl_shadow {
         $($impl)* {
             type WithoutGenerics = <$s as $crate::TS>::WithoutGenerics;
             type OptionInnerType = <$s as $crate::TS>::OptionInnerType;
-            fn ident() -> String { <$s as $crate::TS>::ident() }
-            fn name() -> String { <$s as $crate::TS>::name() }
-            fn inline() -> String { <$s as $crate::TS>::inline() }
-            fn inline_flattened() -> String { <$s as $crate::TS>::inline_flattened() }
+            fn ident(cfg: &$crate::Config) -> String { <$s as $crate::TS>::ident(cfg) }
+            fn name(cfg: &$crate::Config) -> String { <$s as $crate::TS>::name(cfg) }
+            fn inline(cfg: &$crate::Config) -> String { <$s as $crate::TS>::inline(cfg) }
+            fn inline_flattened(cfg: &$crate::Config) -> String { <$s as $crate::TS>::inline_flattened(cfg) }
             fn visit_dependencies(v: &mut impl $crate::TypeVisitor)
             where
                 Self: 'static,
@@ -776,8 +856,8 @@ macro_rules! impl_shadow {
             {
                 <$s as $crate::TS>::visit_generics(v);
             }
-            fn decl() -> String { <$s as $crate::TS>::decl() }
-            fn decl_concrete() -> String { <$s as $crate::TS>::decl_concrete() }
+            fn decl(cfg: &$crate::Config) -> String { <$s as $crate::TS>::decl(cfg) }
+            fn decl_concrete(cfg: &$crate::Config) -> String { <$s as $crate::TS>::decl_concrete(cfg) }
             fn output_path() -> Option<std::path::PathBuf> { <$s as $crate::TS>::output_path() }
         }
     };
@@ -788,12 +868,12 @@ impl<T: TS> TS for Option<T> {
     type OptionInnerType = T;
     const IS_OPTION: bool = true;
 
-    fn name() -> String {
-        format!("{} | null", <T as crate::TS>::name())
+    fn name(cfg: &Config) -> String {
+        format!("{} | null", T::name(cfg))
     }
 
-    fn inline() -> String {
-        format!("{} | null", <T as crate::TS>::inline())
+    fn inline(cfg: &Config) -> String {
+        format!("{} | null", T::inline(cfg))
     }
 
     fn visit_dependencies(v: &mut impl TypeVisitor)
@@ -810,37 +890,21 @@ impl<T: TS> TS for Option<T> {
         <T as crate::TS>::visit_generics(v);
         v.visit::<T>();
     }
-
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
-    }
 }
 
 impl<T: TS, E: TS> TS for Result<T, E> {
     type WithoutGenerics = Result<Dummy, Dummy>;
     type OptionInnerType = Self;
 
-    fn name() -> String {
-        format!(
-            "{{ Ok : {} }} | {{ Err : {} }}",
-            <T as crate::TS>::name(),
-            <E as crate::TS>::name()
-        )
+    fn name(cfg: &Config) -> String {
+        format!("{{ Ok : {} }} | {{ Err : {} }}", T::name(cfg), E::name(cfg))
     }
 
-    fn inline() -> String {
+    fn inline(cfg: &Config) -> String {
         format!(
             "{{ Ok : {} }} | {{ Err : {} }}",
-            <T as crate::TS>::inline(),
-            <E as crate::TS>::inline()
+            T::inline(cfg),
+            E::inline(cfg)
         )
     }
 
@@ -861,34 +925,22 @@ impl<T: TS, E: TS> TS for Result<T, E> {
         <E as crate::TS>::visit_generics(v);
         v.visit::<E>();
     }
-
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
-    }
 }
 
 impl<T: TS> TS for Vec<T> {
     type WithoutGenerics = Vec<Dummy>;
     type OptionInnerType = Self;
 
-    fn ident() -> String {
+    fn ident(_: &Config) -> String {
         "Array".to_owned()
     }
 
-    fn name() -> String {
-        format!("Array<{}>", <T as crate::TS>::name())
+    fn name(cfg: &Config) -> String {
+        format!("Array<{}>", T::name(cfg))
     }
 
-    fn inline() -> String {
-        format!("Array<{}>", <T as crate::TS>::inline())
+    fn inline(cfg: &Config) -> String {
+        format!("Array<{}>", T::inline(cfg))
     }
 
     fn visit_dependencies(v: &mut impl TypeVisitor)
@@ -905,49 +957,35 @@ impl<T: TS> TS for Vec<T> {
         <T as crate::TS>::visit_generics(v);
         v.visit::<T>();
     }
-
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
-    }
 }
 
-// Arrays longer than this limit will be emitted as Array<T>
-const ARRAY_TUPLE_LIMIT: usize = 64;
 impl<T: TS, const N: usize> TS for [T; N] {
     type WithoutGenerics = [Dummy; N];
     type OptionInnerType = Self;
 
-    fn name() -> String {
-        if N > ARRAY_TUPLE_LIMIT {
-            return <Vec<T> as crate::TS>::name();
+    fn name(cfg: &Config) -> String {
+        if N > cfg.array_tuple_limit() {
+            return <Vec<T> as crate::TS>::name(cfg);
         }
 
         format!(
             "[{}]",
             (0..N)
-                .map(|_| <T as crate::TS>::name())
+                .map(|_| T::name(cfg))
                 .collect::<Box<[_]>>()
                 .join(", ")
         )
     }
 
-    fn inline() -> String {
-        if N > ARRAY_TUPLE_LIMIT {
-            return <Vec<T> as crate::TS>::inline();
+    fn inline(cfg: &Config) -> String {
+        if N > cfg.array_tuple_limit() {
+            return <Vec<T> as crate::TS>::inline(cfg);
         }
 
         format!(
             "[{}]",
             (0..N)
-                .map(|_| <T as crate::TS>::inline())
+                .map(|_| T::inline(cfg))
                 .collect::<Box<[_]>>()
                 .join(", ")
         )
@@ -966,18 +1004,6 @@ impl<T: TS, const N: usize> TS for [T; N] {
     {
         <T as crate::TS>::visit_generics(v);
         v.visit::<T>();
-    }
-
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
     }
 }
 
@@ -985,23 +1011,27 @@ impl<K: TS, V: TS, H> TS for HashMap<K, V, H> {
     type WithoutGenerics = HashMap<Dummy, Dummy>;
     type OptionInnerType = Self;
 
-    fn ident() -> String {
+    fn ident(_: &Config) -> String {
         panic!()
     }
 
-    fn name() -> String {
+    fn name(cfg: &Config) -> String {
+        let optional = K::IS_ENUM || cfg.use_v11_hashmap;
         format!(
-            "{{ [key in {}]?: {} }}",
-            <K as crate::TS>::name(),
-            <V as crate::TS>::name()
+            "{{ [key in {}]{}: {} }}",
+            K::name(cfg),
+            if optional { "?" } else { "" },
+            V::name(cfg),
         )
     }
 
-    fn inline() -> String {
+    fn inline(cfg: &Config) -> String {
+        let optional = K::IS_ENUM || cfg.use_v11_hashmap;
         format!(
-            "{{ [key in {}]?: {} }}",
-            <K as crate::TS>::inline(),
-            <V as crate::TS>::inline()
+            "{{ [key in {}]{}: {} }}",
+            K::inline(cfg),
+            if optional { "?" } else { "" },
+            V::inline(cfg),
         )
     }
 
@@ -1009,78 +1039,52 @@ impl<K: TS, V: TS, H> TS for HashMap<K, V, H> {
     where
         Self: 'static,
     {
-        <K as crate::TS>::visit_dependencies(v);
-        <V as crate::TS>::visit_dependencies(v);
+        K::visit_dependencies(v);
+        V::visit_dependencies(v);
     }
 
     fn visit_generics(v: &mut impl TypeVisitor)
     where
         Self: 'static,
     {
-        <K as crate::TS>::visit_generics(v);
+        K::visit_generics(v);
         v.visit::<K>();
-        <V as crate::TS>::visit_generics(v);
+        V::visit_generics(v);
         v.visit::<V>();
     }
 
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        format!(
-            "({{ [key in {}]?: {} }})",
-            <K as crate::TS>::inline(),
-            <V as crate::TS>::inline()
-        )
+    fn inline_flattened(cfg: &Config) -> String {
+        format!("({})", Self::inline(cfg))
     }
 }
 
+// TODO: replace manual impl with dummy struct & `impl_shadow` (like for `JsonValue`)
 impl<I: TS> TS for Range<I> {
     type WithoutGenerics = Range<Dummy>;
     type OptionInnerType = Self;
 
-    fn name() -> String {
-        format!(
-            "{{ start: {}, end: {}, }}",
-            <I as crate::TS>::name(),
-            <I as crate::TS>::name()
-        )
+    fn name(cfg: &Config) -> String {
+        let name = I::name(cfg);
+        format!("{{ start: {name}, end: {name}, }}")
     }
 
     fn visit_dependencies(v: &mut impl TypeVisitor)
     where
         Self: 'static,
     {
-        <I as crate::TS>::visit_dependencies(v);
+        I::visit_dependencies(v);
     }
 
     fn visit_generics(v: &mut impl TypeVisitor)
     where
         Self: 'static,
     {
-        <I as crate::TS>::visit_generics(v);
+        I::visit_generics(v);
         v.visit::<I>();
     }
 
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline() -> String {
-        panic!("{} cannot be inlined", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
+    fn inline(cfg: &Config) -> String {
+        panic!("{} cannot be inlined", Self::name(cfg))
     }
 }
 
@@ -1154,26 +1158,29 @@ mod bytes {
     impl_shadow!(as Vec<u8>: impl TS for bytes::BytesMut);
 }
 
-static LARGE_INT_BINDING: OnceLock<String> = OnceLock::new();
-
 impl_primitives! {
     u8, i8, NonZeroU8, NonZeroI8,
     u16, i16, NonZeroU16, NonZeroI16,
     u32, i32, NonZeroU32, NonZeroI32,
     usize, isize, NonZeroUsize, NonZeroIsize, f32, f64 => "number",
-    u64, i64, NonZeroU64, NonZeroI64,
-    u128, i128, NonZeroU128, NonZeroI128 => LARGE_INT_BINDING
-        .get_or_init(|| std::env::var("TS_RS_LARGE_INT").unwrap_or_else(|_| "bigint".to_owned())),
     bool => "boolean",
     char, Path, PathBuf, String, str,
     Ipv4Addr, Ipv6Addr, IpAddr, SocketAddrV4, SocketAddrV6, SocketAddr => "string",
     () => "null"
 }
 
+impl_large_integers! {
+    u64, i64, NonZeroU64, NonZeroI64,
+    u128, i128, NonZeroU128, NonZeroI128
+}
+
+#[allow(unused_imports)]
 #[rustfmt::skip]
 pub(crate) use impl_primitives;
+#[allow(unused_imports)]
 #[rustfmt::skip]
 pub(crate) use impl_shadow;
+#[allow(unused_imports)]
 #[rustfmt::skip]
 pub(crate) use impl_wrapper;
 
@@ -1191,24 +1198,12 @@ impl TS for Dummy {
     type WithoutGenerics = Self;
     type OptionInnerType = Self;
 
-    fn name() -> String {
+    fn name(_: &Config) -> String {
         "Dummy".to_owned()
     }
 
-    fn decl() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn decl_concrete() -> String {
-        panic!("{} cannot be declared", <Self as crate::TS>::name())
-    }
-
-    fn inline() -> String {
-        panic!("{} cannot be inlined", <Self as crate::TS>::name())
-    }
-
-    fn inline_flattened() -> String {
-        panic!("{} cannot be flattened", <Self as crate::TS>::name())
+    fn inline(cfg: &Config) -> String {
+        panic!("{} cannot be inlined", Self::name(cfg))
     }
 }