+++ ts-rs/tests/integration/chrono.rs
@@ -5,7 +5,7 @@ use chrono::{
     Date, DateTime, Duration, FixedOffset, Local, Month, NaiveDate, NaiveDateTime, NaiveTime, Utc,
     Weekday,
 };
-use ts_rs::TS;
+use ts_rs::{Config, TS};
 
 #[derive(TS)]
 #[ts(export, export_to = "chrono/")]
@@ -25,8 +25,9 @@ struct Chrono {
 
 #[test]
 fn chrono() {
+    let cfg = Config::from_env();
     assert_eq!(
-        Chrono::decl(),
+        Chrono::decl(&cfg),
         "type Chrono = { date: [string, string, string, string], time: string, date_time: [string, string, string, string], duration: [number, number], month: string, weekday: string, };"
     );
 }