+++ ts-rs/tests/integration/list.rs
@@ -1,5 +1,5 @@
 #![allow(dead_code)]
-use ts_rs::TS;
+use ts_rs::{Config, TS};
 
 #[derive(TS)]
 #[ts(export, export_to = "list/")]
@@ -9,5 +9,6 @@ struct List {
 
 #[test]
 fn list() {
-    assert_eq!(List::decl(), "type List = { data: Array<number> | null, };");
+    let cfg = Config::from_env();
+    assert_eq!(List::decl(&cfg), "type List = { data: Array<number> | null, };");
 }